pyproject.toml
[project]
name = "mosaiqjob"
version = "0.2.0"
description = "AI-powered job application document automation pipeline"
requires-python = ">=3.10"

dependencies = [
    "weasyprint",
    "jinja2",
    "pydantic>=2.0",
    "python-dotenv",
    "rich",
    "requests",
    "google-api-python-client",
    "google-auth",
]

[project.scripts]
mosaiq = "app:main"

[project.optional-dependencies]
test = [
    "pytest",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"
